# perf(filters): parallelize feed post-processing