I realized the post processing of the fetched html page is a bottleneck for full_text filter. After profiling, I found:

- The post processing step in dev mode can take 100ms-400ms.
- Most of the computation were hidden inside `tendril` and `html5ever` libraries, which is not easily optimized from this crate.
- Even worse, the post-processing for each feed item was not parallelized. So the latency add up.

So in this patch I moved the compute intensive part into separate tokio tasks so they can be computed at the same time.

For reference, before the change, the sample hackernews full text endpoint [1] takes around 9s to finish. With the request caching, it still takes around 4s. After the change, the uncached version takes 4s and cached version only take less than 1s (on my 8-core machine).

[1]:

``` yaml
  - path: /hackernews.xml
    source: https://news.ycombinator.com/rss
    filters:
      - full_text:
          simplify: true
          append_mode: true
```